Added support for latest release of prompt_toolkit
Version 2 of prompt toolkit made major changes to much of the code
meaning that many of the imports and functions in cbrcli needed to be
updated. Note that this version is not compatible with older versions of
prompt_toolkit, so you will need to upgrade prompt_toolkit if upgrading
to this version of cbrcli. If you are installing from scratch then there
should be no issues as the latest version will automatically be
installed with pip.